+++ orders/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from customer import views
+from orders import views
 
 urlpatterns = [
     path("create/", views.CreateOrderView.as_view(), name="create_order"),
